move nodeGlobalThis to a separate module
