+++ ext/node/01_node.js
@@ -4,21 +4,11 @@
 
 const internals = globalThis.__bootstrap.internals;
 const primordials = globalThis.__bootstrap.primordials;
-const {
-  ArrayPrototypePush,
-  ArrayPrototypeFilter,
-  ObjectDefineProperty,
-  Proxy,
-  ReflectDefineProperty,
-  ReflectDeleteProperty,
-  ReflectGet,
-  ReflectGetOwnPropertyDescriptor,
-  ReflectHas,
-  ReflectOwnKeys,
-  ReflectSet,
-  Set,
-  SetPrototypeHas,
-} = primordials;
+const { ObjectDefineProperty } = primordials;
+import {
+  nodeGlobals,
+  nodeGlobalThis,
+} from "ext:deno_node_loading/00_globals.js";
 
 function assert(cond) {
   if (!cond) {
@@ -27,57 +17,6 @@ function assert(cond) {
 }
 
 let initialized = false;
-const nodeGlobals = {};
-const nodeGlobalThis = new Proxy(globalThis, {
-  get(target, prop) {
-    if (ReflectHas(nodeGlobals, prop)) {
-      return ReflectGet(nodeGlobals, prop);
-    } else {
-      return ReflectGet(target, prop);
-    }
-  },
-  set(target, prop, value) {
-    if (ReflectHas(nodeGlobals, prop)) {
-      return ReflectSet(nodeGlobals, prop, value);
-    } else {
-      return ReflectSet(target, prop, value);
-    }
-  },
-  has(target, prop) {
-    return ReflectHas(nodeGlobals, prop) || ReflectHas(target, prop);
-  },
-  deleteProperty(target, prop) {
-    const nodeDeleted = ReflectDeleteProperty(nodeGlobals, prop);
-    const targetDeleted = ReflectDeleteProperty(target, prop);
-    return nodeDeleted || targetDeleted;
-  },
-  ownKeys(target) {
-    const targetKeys = ReflectOwnKeys(target);
-    const nodeGlobalsKeys = ReflectOwnKeys(nodeGlobals);
-    const nodeGlobalsKeySet = new Set(nodeGlobalsKeys);
-    return [
-      ...ArrayPrototypeFilter(
-        targetKeys,
-        (k) => !SetPrototypeHas(nodeGlobalsKeySet, k),
-      ),
-      ...nodeGlobalsKeys,
-    ];
-  },
-  defineProperty(target, prop, desc) {
-    if (ReflectHas(nodeGlobals, prop)) {
-      return ReflectDefineProperty(nodeGlobals, prop, desc);
-    } else {
-      return ReflectDefineProperty(target, prop, desc);
-    }
-  },
-  getOwnPropertyDescriptor(target, prop) {
-    if (ReflectHas(nodeGlobals, prop)) {
-      return ReflectGetOwnPropertyDescriptor(nodeGlobals, prop);
-    } else {
-      return ReflectGetOwnPropertyDescriptor(target, prop);
-    }
-  },
-});
 
 function initialize(
   nodeModules,
